+++ src/main/java/net/tttprojekt/installer/Main.java
@@ -1,4 +1,4 @@
-package net.tttprojekt.cascademode;
+package net.tttprojekt.installer;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
