+++ src/main/java/net/tttprojekt/installer/download/DownloadTask.java
@@ -1,4 +1,4 @@
-package net.tttprojekt.cascademode.download;
+package net.tttprojekt.installer.download;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
