+++ src/main/java/net/tttprojekt/installer/installer/ForgeInstaller.java
@@ -1,9 +1,9 @@
-package net.tttprojekt.cascademode.installer;
+package net.tttprojekt.installer.installer;
 
-import net.tttprojekt.cascademode.download.Download;
-import net.tttprojekt.cascademode.download.DownloadTaskManager;
-import net.tttprojekt.cascademode.utils.FileDestination;
-import net.tttprojekt.cascademode.utils.ProcessUtils;
+import net.tttprojekt.installer.download.Download;
+import net.tttprojekt.installer.download.DownloadTaskManager;
+import net.tttprojekt.installer.utils.FileDestination;
+import net.tttprojekt.installer.utils.ProcessUtils;
 import org.apache.commons.io.FileUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
