+++ src/main/java/net/tttprojekt/installer/utils/MinecraftChecker.java
@@ -1,4 +1,4 @@
-package net.tttprojekt.cascademode.utils;
+package net.tttprojekt.installer.utils;
 
 import java.util.List;
