+++ src/main/java/net/tttprojekt/installer/utils/OptiFineVersion.java
@@ -1,4 +1,4 @@
-package net.tttprojekt.cascademode.utils;
+package net.tttprojekt.installer.utils;
 
 import lombok.Getter;
